requirements.txt
Djangorestframework==3.14.0
Pandas==1.3.3
Joblib==1.1.0
